Moves contract to launcher package.
